brightbridgeDir/requirements.txt
google-cloud-aiplatform>=1.49.0
google-adk>=1.0.0
python-dotenv==1.1.0
google-generativeai>=0.3.0
requests>=2.31.0